Centipede: remove annoying unnecessary message
It's the only message runner prints.
And it's printed on each process start repeatedly.
And compiling runner with sanitizers is the intended more.
And nobody sees it anyway since Centipede discards the output.
Remove it.

If it's really necessary, this condition should be detected
during binary assessment on start and printed once by Centipede
binary rather than the runner.